// src/types/estimate-details.types.ts
type ProjectManagementValues = {
  status: SelectOption | null
  type: SelectOption | null
  woNumber: string | null
  poNumber: string | null
  projectName: string | null
  woaStartDate: string | null
  woaCompletionDate: string | null
  clientStartDate: string | null
  clientDueDate: string | null
  clientWalkthroughDate: string | null
  clientSignOffDate: string | null
  overrideProjectStatus: SelectOption | null
  amount?: string | null
  estimateDocument?: any
}

type ProjectInvoicingAndPaymentFormValues = {
  originalSOWAmount: number
  sowLink: string | null
  invoiceLink: string | null
  finalSOWAmount: number
  invoiceNumber: string | null
  invoiceAttachment: File | null
  invoiceBackDate: string | null
  paymentTerms: SelectOption | null
  woaInvoiceDate: string | null
  woaExpectedPayDate: string | null
  overPayment: number | null
  remainingPayment: number | null
  payment: number | null
}

type ContactsFormValues = {
  projectCoordinator: SelectOption | null
  projectCoordinatorPhoneNumber: string | null
  projectCoordinatorExtension: string | null
  fieldProjectManager: SelectOption | null
  fieldProjectManagerPhoneNumber: string | null
  fieldProjectManagerExtension: string | null
  superName: string | null
  superPhoneNumber: string | null
  superPhoneNumberExtension: string | null
  superEmail: string | null
  client: SelectOption | null
}

type ProjectDetailsLocationFormValues = {
  address: string | null
  city: string | null
  state: SelectOption | null
  zip: string | null
  market: SelectOption | null
  gateCode: string | null
  lockBoxCode: string | null
  hoaContactPhoneNumber: string | null
  hoaContactExtension: string | null
  hoaContactEmail: string | null
}

type ProjectDetailsMiscFormValues = {
  dateCreated: string | null
  activeDate: string | null
  punchDate: string | null
  closedDate: string | null
  clientPaidDate: string | null
  collectionDate: string | null
  disputedDate: string | null
  woaPaidDate: string | null
  dueDateVariance: number | null
  payDateVariance: number | null
  payVariance: number | null
}

// Project Details form values types
export type ProjectDetailsFormValues = ProjectManagementValues &
  ProjectInvoicingAndPaymentFormValues &
  ContactsFormValues &
  ProjectDetailsLocationFormValues &
  ProjectDetailsMiscFormValues